15.7.15 Fix minor bugs (http://www.eonet.ne.jp/~gakana/tablacus/history/)

+++ Debug/script/index.js
@@ -821,14 +821,16 @@ te.OnKeyMessage = function (Ctrl, hwnd, msg, key, keydata)
 			default:
 				if (window.g_menu_click) {
 					if (key == VK_RETURN) {
-						var hSubMenu = api.GetSubMenu(window.g_menu_handle, window.g_menu_pos);
-						if (hSubMenu) {
-							var mii = api.Memory("MENUITEMINFO");
-							mii.cbSize = mii.Size;
-							mii.fMask = MIIM_SUBMENU;
-							api.SetMenuItemInfo(window.g_menu_handle, window.g_menu_pos, true, mii);
-							api.DestroyMenu(hSubMenu);
-							api.PostMessage(hwnd, WM_CHAR, VK_LBUTTON, 0);
+						if (window.g_menu_click === true) {
+							var hSubMenu = api.GetSubMenu(window.g_menu_handle, window.g_menu_pos);
+							if (hSubMenu) {
+								var mii = api.Memory("MENUITEMINFO");
+								mii.cbSize = mii.Size;
+								mii.fMask = MIIM_SUBMENU;
+								api.SetMenuItemInfo(window.g_menu_handle, window.g_menu_pos, true, mii);
+								api.DestroyMenu(hSubMenu);
+								api.PostMessage(hwnd, WM_CHAR, VK_LBUTTON, 0);
+							}
 						}
 						window.g_menu_button = api.GetKeyState(VK_SHIFT) >= 0 ? 1 : 2;
 						if (api.GetKeyState(VK_CONTROL) < 0) {
@@ -2107,7 +2109,7 @@ g_mouse =
 					if (window.g_menu_click != 2) {
 						window.g_menu_button = str;
 					}
-					if (window.g_menu_click) {
+					if (window.g_menu_click === true) {
 						var hSubMenu = api.GetSubMenu(window.g_menu_handle, window.g_menu_pos);
 						if (hSubMenu) {
 							var mii = api.Memory("MENUITEMINFO");
